+++ gradle.properties
@@ -16,6 +16,6 @@ org.gradle.jvmargs=-Xmx1536m
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 
-SAMPLE_VERSION_NAME=0.0.13
+SAMPLE_VERSION_NAME=0.0.15
 
 KIN_ECOSYSTEM_SDK_VERSION_NAME=0.0.4
