+++ kin-ecosystem-sdk/src/main/java/com/kin/ecosystem/data/blockchain/BlockchainSource.java
@@ -32,6 +32,7 @@ public class BlockchainSource implements IBlockchainSource {
 
     private static volatile BlockchainSource instance;
 
+    private static final String ISSUER_PRIVATE_TEST = "GBQ3DQOA7NF52FVV7ES3CR3ZMHUEY4LTHDAQKDTO6S546JCLFPEQGCPK";
     private KinClient kinClient;
     private KinAccount account;
     private ObservableData<Integer> balance = ObservableData.create(0);
@@ -58,7 +59,7 @@ public class BlockchainSource implements IBlockchainSource {
 
     private BlockchainSource(@NonNull final Context context)
         throws InitializeException {
-        this.kinClient = new KinClient(context, StellarNetwork.NETWORK_TEST.getProvider());
+        this.kinClient = new KinClient(context, Network.NETWORK_PRIVATE_TEST.getProviderForIssuer(ISSUER_PRIVATE_TEST));
         createKinAccountIfNeeded();
         getCurrentBalance();
     }
